Rename sythetic_event_factories
