+++ lib/react_client/utils.dart
@@ -4,7 +4,7 @@ import 'package:react/react_client/js_backed_map.dart';
 import 'package:react/react_client/react_interop.dart';
 
 import 'package:react/src/react_client/event_prop_key_to_event_factory.dart';
-import 'package:react/src/react_client/synthetic_event_factories.dart';
+import 'package:react/src/react_client/event_factory.dart';
 
 /// Prepares [children] to be passed to the ReactJS [React.createElement] and
 /// the Dart [react.Component].
