+++ lib/src/react_client/event_factory.dart
